Move location interface into a file
