+++ src/app/components/map-search/map-search.component.html
@@ -17,7 +17,7 @@
   <div
     [@slideInOut]="isSearchVisible ? 'in' : 'out'"
     [ngStyle]="{ 'height': isSearchVisible ? 'auto' : '0', 'overflow': isSearchVisible ? 'visible' : 'hidden' }"
-    class="w-full space-y-2"
+    class="w-full space-y-2 z-40"
   >
     <div class="relative">
       <input
