+++ src/app/interfaces/location.interface.ts
@@ -1,0 +1,7 @@
+export interface Location {
+  title: string;
+  address: string;
+  coordinates: google.maps.LatLng;
+  image?: string;
+  marker?: google.maps.Marker;
+}
